feat: add address sub-components
